Split up the controls further to allow testing
